Js added to populate add_new and tweaked admin js
