+++ templates/admin.html
@@ -1,5 +1,5 @@
 <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
-<script type="text/javascript" src="{{url_for('static', filename='/js/categories.js')}}"></script>
+<script type="text/javascript" src="{{url_for('static', filename='/js/admin_categories.js')}}"></script>
 {% extends "layout.html" %}
 {% block content %}
     {% if is_admin %}
